CosmoCat_C#ScriptsV2
C# scripts for Unity. Blast controller is to be attached to a bullet prefab. Player controller allows the player to shoot through instantiating a prefab and do a charged shot with variable damage and size based on how long the Fire1 button is held down. Bounds the player within the camera and move along the axis and arrow inputs at speed.